moving loggers.py and error.py to tools directory
