+++ parcels/kernels/__init__.py
@@ -1,3 +1,2 @@
 from .advection import *  # noqa
 from .diffusion import *  # noqa
-from .error import *  # noqa
