+++ parcels/particle.py
@@ -1,6 +1,6 @@
-from parcels.kernels.error import ErrorCode
+from parcels.tools.error import ErrorCode
 from parcels.field import Field
-from parcels.loggers import logger
+from parcels.tools.loggers import logger
 from operator import attrgetter
 import numpy as np
 from ctypes import c_void_p
