+++ parcels/tools/__init__.py
@@ -1,2 +1,4 @@
 from .timer import *  # noqa
 from .converters import *  # noqa
+from .loggers import *  # noqa
+from .error import *  # noqa
